requirements.txt
langchain 
pymongo 
bs4
openai
tiktoken
gradio
requests
lxml 
argparse
unstructured
python-dotenv